### 2022-01-12 #4055 @whedon
[PRE REVIEW]: title: 'terrainr: An R package for creating immersive virtual environments
**Submitting author:** @mikemahoney218 (<a href="http://orcid.org/0000-0003-2402-304X">Michael Mahoney</a>)
**Repository:** <a href="https://github.com/ropensci/terrainr" target ="_blank">https://github.com/ropensci/terrainr</a>
**Version:** v0.5.1
