### 2022-01-12 #4055 @mikemahoney218
Apologies for the paper not rendering -- the paper lives in https://github.com/ropensci/terrainr/tree/paper which I could specify in the pre-submission rendering service, but I couldn't figure out how to provide the branch to the actual submission workflow.
