### 2022-01-12 6be1235
Remove extraneous bibtex
