+++ paper.bib
@@ -1,126 +1,3 @@
-@article{Dransch2010,
-  author = {Doris Dransch and Patrick Kothur and Sven Schulte and Volker Klemann and Henryk Dobslaw},
-  title = {Assessing the quality of geoscientific simulation models with visual analytics methods – a design study},
-  journal = {International Journal of Geographical Information Science},
-  volume = {24},
-  number = {10},
-  pages = {1459-1479},
-  year  = {2010},
-  publisher = {Taylor & Francis},
-  doi = {10.1080/13658816.2010.510800}
-}
-
-@article{Nakame2001,
-  title = "Rendering of landscapes for environmental assessment",
-  journal = "Landscape and Urban Planning",
-  volume = "54",
-  number = "1",
-  pages = "19 - 32",
-  year = "2001",
-  note = "Our Visual Landscape: analysis, modeling, visualization and protection",
-  issn = "0169-2046",
-  doi = "https://doi.org/10.1016/S0169-2046(01)00123-2",
-  url = "http://www.sciencedirect.com/science/article/pii/S0169204601001232",
-  author = "Eihachiro Nakamae and Xueying Qin and Katsumi Tadamura",
-  keywords = "Photo-realistic images, Panoramic images, Video sequence montages",
-}
-
-@incollection{Gahegan2005,
-  title = "Chapter 4 - Beyond Tools: Visual Support for the Entire Process of GIScience",
-  editor = "Jason Dykes and Alan M. MacEachren and Menno-Jan Kraak",
-  booktitle = "Exploring Geovisualization",
-  publisher = "Elsevier",
-  address = "Oxford",
-  pages = "83 - 99",
-  year = "2005",
-  series = "International Cartographic Association",
-  isbn = "978-0-08-044531-1",
-  doi = "https://doi.org/10.1016/B978-008044531-1/50422-X",
-  url = "http://www.sciencedirect.com/science/article/pii/B978008044531150422X",
-  author = "Mark Gahegan",
-  abstract = "Publisher Summary
-This chapter presents an overview of the science process, illustrates examples of the way current approaches to visualization support certain research activities, and suggests a framework by which additional aspects of: (1) science activity, (2) reasoning mode (inference), (3) outcomes, and (4) pre-requisite data and knowledge can be used to better define the role or roles that different visualization tools can play. To bring visualization tools to the fore, it is important to understand clearly the roles that they play within the entire process of scientific investigation including the various tasks that a researcher might undertake and the different kinds of reasoning that the researcher wish visualization to support or enable. The research challenges that are associated with the use and integration of geovisualization tools from a more semantic, process-oriented approach are also summarized. As geovisualization develops, a broader vision is required by which relevant tools and techniques might be orchestrated into something greater than the sum of their parts. This vision is likely to include perspectives from other disciplines and may continue to incorporate techniques and knowledge from cognate fields with expertise in visualization."
-}
-
-@article{Schmid2001,
-  title = "The emerging role of visual resource assessment and visualisation in landscape planning in Switzerland",
-  journal = "Landscape and Urban Planning",
-  volume = "54",
-  number = "1",
-  pages = "213 - 221",
-  year = "2001",
-  note = "Our Visual Landscape: analysis, modeling, visualization and protection",
-  issn = "0169-2046",
-  doi = "https://doi.org/10.1016/S0169-2046(01)00137-2",
-  url = "http://www.sciencedirect.com/science/article/pii/S0169204601001372",
-  author = "Willy A Schmid",
-  keywords = "Spatial planning, Landscape planning, Visual resource, 3D-visualization, Aesthetics, Cultural heritage",
[…]
