+++ paper.bib
@@ -10,7 +10,7 @@
 @manual{TNM,
   author = {{U.S. Geological Survey National Geospatial Program}},
   year = {2020},
-  title = {The National Map},
+  title = {{The National Map}},
   language = {English},
   url = {https://viewer.nationalmap.gov/services/},
   address = {Washington D. C., United States of America},
@@ -43,14 +43,15 @@
   volume          = {71},
   year            = {2021},
   pages           = {53-65},
+  doi             = {10.1007/s42489-020-00069-6},
 }
 
-@unpublished{vrsPaper,
-  author  = {Mahoney, Michael J and Beier, Colin M and Ackerman, Aidan C},
-  title = {Interactive landscape simulations for visual resource assessment},
-  note = {In Press},
-  journal = {Proceedings of the 2021 Visual Resources Stewardship Conference},
-  year = {2021}
+@inproceedings{vrsPaper,
+  author          = {Mahoney, Michael J and Beier, Colin M and Ackerman, Aidan C},
+  booktitle       = {VRSC 2021 Conference Proceedings},
+  editor          = {Hoffman, Robin and Chamberlain, Brent and Smardon, Richard},
+  title           = {Interactive landscape simulations for visual resource assessment},
+  year            = {2021}
 }
 
 @article{Hruby2019,
