### 2022-01-12 #4055 @mikemahoney218
I believe that should address the reference check! Let me know if I missed anything.
