### 2022-01-13 #4060 @danielskatz
@mikemahoney218 - the zenodo archive has a different title and authors than the paper. Typically, these are the same for a JOSS-related paper.  Note that you can manually change the metadata of the zenodo archive.
