### 2022-01-13 #4060 @danielskatz
Also, in the paper, please change "section 3 of" to "Section 3 of"
