### 2022-01-13 #4060 @mikemahoney218
@danielskatz Thanks! I believe I've fixed both issues.
