### 2022-01-13 #4060 @danielskatz
@arfon - can you see what's happened here?  I think this has large images that make things slow, but not this slow.
